// lib/openapi-diff/spec-differ/diff-finder/difference.d.ts
import {
    DiffResultAction,
    DiffResultCode,
    DiffResultEntity,
    DiffResultSource,
    DiffResultSpecEntityDetails
} from '../../../api-types';

interface Difference {
    action: DiffResultAction;
    code: DiffResultCode;
    entity: DiffResultEntity;
    sourceSpecEntityDetails: DiffResultSpecEntityDetails[];
    destinationSpecEntityDetails: DiffResultSpecEntityDetails[];
    details?: any;
}